app/requirements.txt
fastapi==0.87.0
python-multipart==0.0.9
tensorflow==2.13.1
keras==2.13.1
pandas==1.3.5
numpy==1.24.1
uvicorn==0.19.0
pydantic==1.10.15
pillow==9.0.1